Made some UI changes and fixed some interaction and typography bugs💯
