+++ app/company/aboutus/page.tsx
@@ -7,7 +7,7 @@ import WhatWeOffer from "@/components/About-Page/WhatWeOffer";
 const Aboutus = () => {
   return (
     <main className="mx-auto max-w-[1440px] px-6 pt-24 xl:px-8">
-      <h1 className="mb-12 text-6xl text-center md:text-7xl xl:text-9xl">
+      <h1 className="mb-12 text-center text-6xl md:text-7xl xl:text-9xl">
         Agency<span className="text-brandDark">.io</span>
       </h1>
       <div className="space-y-40">
