+++ app/not-found.tsx
@@ -30,7 +30,7 @@ const NotFound = () => {
           </p>
 
           <div className="flex items-center gap-6 md:gap-10 xs:gap-4">
-            <Button className="gap-1 px-8 text-lg" variant={"plain"}>
+            <Button className="gap-1 px-8 text-lg" variant={"plain"} asChild>
               <Link href={`/?splashed=true`} prefetch={false}>
                 Home
               </Link>
