+++ components/Home-Page/Search.tsx
@@ -35,8 +35,8 @@ const Search = () => {
 
   return (
     <section className="flex w-full p-4 pb-0 lg:max-h-[900px] lg:gap-8 lg:p-8 lg:pb-0 xl:h-screen">
-      <div className="items-center justify-end hidden xl:flex xl:basis-1/2 ">
-        <div className="relative w-full overflow-hidden h-4/5 rounded-xl">
+      <div className="hidden items-center justify-end xl:flex xl:basis-1/2 ">
+        <div className="relative h-4/5 w-full overflow-hidden rounded-xl">
           <Image
             src={SearchImg}
             placeholder="blur"
@@ -47,11 +47,11 @@ const Search = () => {
           />
         </div>
       </div>
-      <div className="flex flex-col items-center w-full gap-8 pt-16 md:pt-28 xl:basis-1/2 xl:items-start xl:justify-center xl:pt-0">
-        <div className="flex flex-wrap justify-center max-w-3xl gap-5 overflow-scroll activities dark:text-white xl:justify-start">
+      <div className="flex w-full flex-col items-center gap-8 pt-16 md:pt-28 xl:basis-1/2 xl:items-start xl:justify-center xl:pt-0">
+        <div className="activities flex max-w-3xl flex-wrap justify-center gap-5 overflow-scroll dark:text-white xl:justify-start">
           <Button
             variant={"outline"}
-            className="px-6 py-2 text-white border-black bg-background dark:border-white dark:bg-white dark:text-black"
+            className="border-black bg-background px-6 py-2 text-white dark:border-white dark:bg-white dark:text-black"
           >
             All
           </Button>
@@ -66,23 +66,23 @@ const Search = () => {
             );
           })}
 
-          <NBgLink prompt="View more" href="#" />
+          <NBgLink prompt="View more" href="/services" />
         </div>
 
         {user === undefined ? (
           <div className="h-12 w-80 animate-pulse bg-gray-400 md:w-[600px]"></div>
         ) : user ? (
-          <h1 className="text-[26px] dark:text-white md:text-5xl">
+          <p className="text-center text-[26px] dark:text-white md:text-5xl">
             What&apos;s the{" "}
             <span className="text-brandDark dark:text-brandLight">plan</span>
             {user.displayName ? ", " + user.displayName.split(" ")[0] : ""}?
-          </h1>
+          </p>
         ) : (
-          <h1 className="text-2xl dark:text-white md:text-5xl">
+          <p className="text-center text-2xl dark:text-white md:text-5xl">
             Discover your{" "}
             <span className="text-brandDark dark:text-brandLight">next</span>{" "}
             adventure
-          </h1>
+          </p>
         )}
 
         <div className="flex flex-wrap gap-4">
@@ -102,7 +102,7 @@ const Search = () => {
             >
               <span className="md:hidden">
                 {button.plan === "generate"
-                  ? "Build w/A.I"
+                  ? "A.I Gen"
                   : button.text.split(" ")[0]}
               </span>
               <span className="hidden md:block">{button.text}</span>
