+++ components/UI/Carousel/SharedPagePlaceCarousel.tsx
@@ -53,8 +53,8 @@ const SharedPagePlaceCarousel = ({
       <LeftArrow />
 
       {items.length === 0 ? (
-        <div className="flex h-[300px] w-full items-center justify-center gap-4 rounded-xl border">
-          <p className="text-xl">
+        <div className="flex h-[300px] w-full items-center justify-center gap-4 rounded-xl border border-black dark:border-white">
+          <p className="w-4/5 text-center text-xl">
             No related content is available yet, check back soon!
           </p>
         </div>
