+++ static-data/destinations.ts
@@ -1,0 +1,612 @@
+import Amsterdam from "@/public/Destinations/Amsterdam.jpg";
+import Australia from "@/public/Destinations/Australia.jpg";
+import Bali from "@/public/Destinations/Bali.jpg";
+import Barcelona from "@/public/Destinations/Barcelona.jpg";
+import Brazil from "@/public/Destinations/Brazil.webp";
+import Greece from "@/public/Destinations/Greece.jpg";
+import Hawaii from "@/public/Destinations/Hawaii.jpg";
+import Iceland from "@/public/Destinations/Iceland.jpg";
+import Italy from "@/public/Destinations/Italy.jpg";
+import Japan from "@/public/Destinations/Japan.webp";
+import KyotoAutumn from "@/public/Destinations/KyotoAutumn.jpg";
+import KyotoSpring from "@/public/Destinations/KyotoSpring.jpg";
+import Mexico from "@/public/Destinations/Mexico.jpg";
+import NewEngland from "@/public/Destinations/NewEngland.jpg";
+import NewZealand from "@/public/Destinations/NewZealand.jpg";
+import Reykjavik from "@/public/Destinations/Reykjavik.jpg";
+import Spain from "@/public/Destinations/Spain.webp";
+import Thailand from "@/public/Destinations/Thailand.webp";
+import Zermatt from "@/public/Destinations/Zermatt.webp";
+
+import Bora from "../public/Destinations/Bora.webp";
+import Machu from "../public/Destinations/Machu.webp";
+import Maldives from "../public/Destinations/Maldives.webp";
+import Santorini from "../public/Destinations/Santorini.webp";
+
+export const locationsByMonth = [
+  {
+    _id: 1,
+    name: "January",
+    slug: "bora-bora",
+    extra: "Bora Bora, French Polynesia",
+    displayImage: { url: Bora, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 2,
+    name: "February",
+    slug: "venice",
+    extra: "Venice, Italy",
+    displayImage: { url: Italy, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 3,
+    name: "March",
+    slug: "cancun",
+    extra: "Cancun, Mexico",
+    displayImage: { url: Mexico, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 4,
+    name: "April",
+    slug: "kyoto",
+    extra: "Kyoto, Japan",
+    displayImage: { url: Japan, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 5,
+    name: "May",
+    slug: "barcelona",
+    extra: "Barcelona, Spain",
+    displayImage: { url: Spain, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 6,
+    name: "June",
+    slug: "santorini",
+    extra: "Santorini, Greece",
+    displayImage: { url: Greece, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 7,
+    name: "July",
+    slug: "maui",
+    extra: "Maui, Hawaii",
+    displayImage: { url: Hawaii, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 8,
+    name: "August",
+    slug: "rio-de-janeiro",
+    extra: "Rio de Janeiro, Brazil",
+    displayImage: { url: Brazil, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 9,
+    name: "September",
+    slug: "machu-picchu",
+    extra: "Machu Picchu, Peru",
+    displayImage: { url: Machu, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 10,
+    name: "October",
+    slug: "maldives",
+    extra: "Maldives",
+    displayImage: { url: Maldives, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 11,
+    name: "November",
+    slug: "new-zealand",
+    extra: "New Zealand",
+    displayImage: { url: NewZealand, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 12,
+    name: "December",
+    slug: "thailand",
+    extra: "Thailand",
+    displayImage: { url: Thailand, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+];
+
+export const locationsBySeason = [
+  {
+    _id: 1,
+    name: "Spring",
+    slug: "kyoto",
+    extra: "Kyoto, Japan",
+    displayImage: { url: KyotoSpring, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 2,
+    name: "Spring",
+    slug: "amsterdam",
+    extra: "Amsterdam, Netherlands",
+    displayImage: { url: Amsterdam, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 3,
+    name: "Summer",
+    slug: "bali",
+    extra: "Bali, Indonesia",
+    displayImage: { url: Bali, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 4,
+    name: "Summer",
+    slug: "barcelona",
+    extra: "Barcelona, Spain",
+    displayImage: { url: Barcelona, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 5,
+    name: "Autumn (Fall)",
+    slug: "kyoto",
+    extra: "Kyoto, Japan",
+    displayImage: { url: KyotoAutumn, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 6,
+    name: "Autumn (Fall)",
+    slug: "new-england",
+    extra: "New England, USA",
+    displayImage: { url: NewEngland, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 7,
+    name: "Winter",
+    slug: "reykjavik",
+    extra: "Reykjavik, Iceland",
+    displayImage: { url: Reykjavik, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 8,
+    name: "Winter",
+    slug: "zermatt",
+    extra: "Zermatt, Switzerland",
+    displayImage: { url: Zermatt, alt: "" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+];
+
+export const recommendedDes = [
+  {
+    _id: 1,
+    name: "Italy",
+    slug: "italy",
+    displayImage: { url: Italy, alt: "Italy" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 2,
+    name: "Japan",
+    slug: "japan",
+    displayImage: { url: Japan, alt: "Japan" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 3,
+    name: "Greece",
+    slug: "greece",
+    displayImage: { url: Greece, alt: "Greece" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 4,
+    name: "New Zealand",
+    slug: "new-zealand",
+    displayImage: { url: NewZealand, alt: "New Zealand" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 5,
+    name: "Thailand",
+    slug: "thailand",
+    displayImage: { url: Thailand, alt: "Thailand" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 6,
+    name: "Iceland",
+    slug: "iceland",
+    displayImage: { url: Iceland, alt: "Iceland" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 7,
+    name: "Australia",
+    slug: "australia",
+    displayImage: { url: Australia, alt: "Australia" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+  {
+    _id: 8,
+    name: "Spain",
+    slug: "spain",
+    displayImage: { url: Spain, alt: "Spain" },
+    country: "",
+    description: "",
+    popularSpots: [],
+    pictures: [],
+    essentials: [],
+    posts: [],
+    whyWeLove: [],
+    coordinates: {
+      altitude: 0,
+      latitude: 0,
+      longitude: 0,
+    },
+    faqs: [],
+  },
+];
+
+export const destinations = [
+  {
+    id: 1,
+    destination: "Maldives",
+    note: "Experience paradise on Earth with pristine white sandy beaches, crystal-clear turquoise waters, and luxurious over-water bungalows.",
+    imgsrc: Maldives,
+  },
+  {
+    id: 2,
+    destination: "Santorini",
+    note: "Indulge in the charm of this Greek island paradise known for its stunning sunsets, whitewashed buildings, and breathtaking views over the Aegean Sea.",
+    imgsrc: Santorini,
+  },
+  {
+    id: 3,
+    destination: "Machu Picchu",
+    note: "Embark on a journey to the ancient ruins of the Incas, hidden amidst the Andes mountains, and marvel at the awe-inspiring beauty and rich history.",
+    imgsrc: Machu,
+  },
+  {
+    id: 4,
+    destination: "Bora Bora",
+    note: "Immerse yourself in the allure of this French Polynesian gem, where turquoise lagoons, vibrant coral reefs, and luxurious over-water bungalows await.",
+    imgsrc: Bora,
+  },
+];
