+++ static-data/images.js
@@ -1,28 +1,3 @@
-import React from "react";
-import Amsterdam from "@/public/Destinations/Amsterdam.jpg";
-import Australia from "@/public/Destinations/Australia.jpg";
-import Bali from "@/public/Destinations/Bali.jpg";
-import Barcelona from "@/public/Destinations/Barcelona.jpg";
-import Brazil from "@/public/Destinations/Brazil.webp";
-import Greece from "@/public/Destinations/Greece.jpg";
-import Hawaii from "@/public/Destinations/Hawaii.jpg";
-import Iceland from "@/public/Destinations/Iceland.jpg";
-import Italy from "@/public/Destinations/Italy.jpg";
-import Japan from "@/public/Destinations/Japan.webp";
-import KyotoAutumn from "@/public/Destinations/KyotoAutumn.jpg";
-import KyotoSpring from "@/public/Destinations/KyotoSpring.jpg";
-import Mexico from "@/public/Destinations/Mexico.jpg";
-import NewEngland from "@/public/Destinations/NewEngland.jpg";
-import NewZealand from "@/public/Destinations/NewZealand.jpg";
-import Reykjavik from "@/public/Destinations/Reykjavik.jpg";
-import Spain from "@/public/Destinations/Spain.webp";
-import Thailand from "@/public/Destinations/Thailand.webp";
-import Zermatt from "@/public/Destinations/Zermatt.webp";
-import Cabin from "@/public/Properties/Cabin.jpg";
-import Castle from "@/public/Properties/Castle.jpg";
-import Cottage from "@/public/Properties/Cottage.jpg";
-import Lodge from "@/public/Properties/Lodge.jpg";
-import Villa from "@/public/Properties/Villa.jpg";
 import {
   FaBicycle,
   FaCamera,
@@ -36,428 +11,24 @@ import {
 } from "react-icons/fa";
 import { GiCampingTent, GiCookingPot, GiMountainCave } from "react-icons/gi";
 
-import Image1 from "../public/Carousel/Image1.webp";
-import Image2 from "../public/Carousel/Image2.webp";
-import Image3 from "../public/Carousel/Image3.webp";
-import Image4 from "../public/Carousel/Image4.webp";
-import Image5 from "../public/Carousel/Image5.webp";
-import Image6 from "../public/Carousel/Image6.webp";
-import Image7 from "../public/Carousel/Image7.webp";
-import Image8 from "../public/Carousel/Image8.jpg";
-import Pic1 from "../public/CTA/Pic1.jpg";
-import Pic2 from "../public/CTA/Pic2.jpg";
-import Pic3 from "../public/CTA/Pic3.jpg";
-import Pic4 from "../public/CTA/Pic4.jpg";
-import Bora from "../public/Destinations/Bora.webp";
-import Machu from "../public/Destinations/Machu.webp";
-import Maldives from "../public/Destinations/Maldives.webp";
-import Santorini from "../public/Destinations/Santorini.webp";
-
-export const locationsByMonth = [
-  {
-    _id: 1,
-    name: "January",
-    slug: "bora-bora",
-    extra: "Bora Bora, French Polynesia",
-    displayImage: { url: Bora, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 2,
-    name: "February",
-    slug: "venice",
-    extra: "Venice, Italy",
-    displayImage: { url: Italy, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 3,
-    name: "March",
-    slug: "cancun",
-    extra: "Cancun, Mexico",
-    displayImage: { url: Mexico, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 4,
-    name: "April",
-    slug: "kyoto",
-    extra: "Kyoto, Japan",
-    displayImage: { url: Japan, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 5,
-    name: "May",
-    slug: "barcelona",
-    extra: "Barcelona, Spain",
-    displayImage: { url: Spain, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 6,
-    name: "June",
-    slug: "santorini",
-    extra: "Santorini, Greece",
-    displayImage: { url: Greece, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 7,
-    name: "July",
-    slug: "maui",
-    extra: "Maui, Hawaii",
-    displayImage: { url: Hawaii, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 8,
-    name: "August",
-    slug: "rio-de-janeiro",
-    extra: "Rio de Janeiro, Brazil",
-    displayImage: { url: Brazil, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 9,
-    name: "September",
-    slug: "machu-picchu",
-    extra: "Machu Picchu, Peru",
-    displayImage: { url: Machu, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 10,
-    name: "October",
-    slug: "maldives",
-    extra: "Maldives",
-    displayImage: { url: Maldives, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 11,
-    name: "November",
-    slug: "new-zealand",
-    extra: "New Zealand",
-    displayImage: { url: NewZealand, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 12,
-    name: "December",
-    slug: "thailand",
-    extra: "Thailand",
-    displayImage: { url: Thailand, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-];
-
-export const locationsBySeason = [
-  {
-    _id: 1,
-    name: "Spring",
-    slug: "kyoto",
-    extra: "Kyoto, Japan",
-    displayImage: { url: KyotoSpring, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 2,
-    name: "Spring",
-    slug: "amsterdam",
-    extra: "Amsterdam, Netherlands",
-    displayImage: { url: Amsterdam, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 3,
-    name: "Summer",
-    slug: "bali",
-    extra: "Bali, Indonesia",
-    displayImage: { url: Bali, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 4,
-    name: "Summer",
-    slug: "barcelona",
-    extra: "Barcelona, Spain",
-    displayImage: { url: Barcelona, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 5,
-    name: "Autumn (Fall)",
-    slug: "kyoto",
-    extra: "Kyoto, Japan",
-    displayImage: { url: KyotoAutumn, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 6,
-    name: "Autumn (Fall)",
-    slug: "new-england",
-    extra: "New England, USA",
-    displayImage: { url: NewEngland, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 7,
-    name: "Winter",
-    slug: "reykjavik",
-    extra: "Reykjavik, Iceland",
-    displayImage: { url: Reykjavik, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 8,
-    name: "Winter",
-    slug: "zermatt",
-    extra: "Zermatt, Switzerland",
-    displayImage: { url: Zermatt, alt: "" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-];
+import VacationImg1 from "../public/Carousel/Image1.webp";
+import VacationImg2 from "../public/Carousel/Image2.webp";
+import VacationImg3 from "../public/Carousel/Image3.webp";
+import VacationImg4 from "../public/Carousel/Image4.webp";
+import VacationImg5 from "../public/Carousel/Image5.webp";
+import VacationImg6 from "../public/Carousel/Image6.webp";
+import VacationImg7 from "../public/Carousel/Image7.webp";
+import VacationImg8 from "../public/Carousel/Image8.jpg";
+import CTAPic1 from "../public/CTA/Pic1.jpg";
+import CTAPic2 from "../public/CTA/Pic2.jpg";
+import CTAPic3 from "../public/CTA/Pic3.jpg";
+import CTAPic4 from "../public/CTA/Pic4.jpg";
+import ExecImg from "../public/Hero/exec.jpg";
+import PropertyImg3 from "../public/Properties/Cabin.jpg";
+import PropertyImg5 from "../public/Properties/Castle.jpg";
+import PropertyImg1 from "../public/Properties/Cottage.jpg";
+import PropertyImg4 from "../public/Properties/Lodge.jpg";
+import PropertyImg2 from "../public/Properties/Villa.jpg";
 
 export const staticExperiencesData = [
   {
@@ -543,204 +114,22 @@ export const staticExperiencesData = [
   },
 ];
 
-export const recommendedDes = [
-  {
-    _id: 1,
-    name: "Italy",
-    slug: "italy",
-    displayImage: { url: Italy, alt: "Italy" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 2,
-    name: "Japan",
-    slug: "japan",
-    displayImage: { url: Japan, alt: "Japan" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 3,
-    name: "Greece",
-    slug: "greece",
-    displayImage: { url: Greece, alt: "Greece" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 4,
-    name: "New Zealand",
-    slug: "new-zealand",
-    displayImage: { url: NewZealand, alt: "New Zealand" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 5,
-    name: "Thailand",
-    slug: "thailand",
-    displayImage: { url: Thailand, alt: "Thailand" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 6,
-    name: "Iceland",
-    slug: "iceland",
-    displayImage: { url: Iceland, alt: "Iceland" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 7,
-    name: "Australia",
-    slug: "australia",
-    displayImage: { url: Australia, alt: "Australia" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-  {
-    _id: 8,
-    name: "Spain",
-    slug: "spain",
-    displayImage: { url: Spain, alt: "Spain" },
-    country: "",
-    description: "",
-    popularSpots: [],
-    pictures: [],
-    essentials: [],
-    posts: [],
-    whyWeLove: [],
-    coordinates: {
-      altitude: 0,
-      latitude: 0,
-      longitude: 0,
-    },
-    faqs: [],
-  },
-];
-
-export const destinations = [
-  {
-    id: 1,
-    destination: "Maldives",
-    note: "Experience paradise on Earth with pristine white sandy beaches, crystal-clear turquoise waters, and luxurious over-water bungalows.",
-    imgsrc: Maldives,
-  },
-  {
-    id: 2,
-    destination: "Santorini",
-    note: "Indulge in the charm of this Greek island paradise known for its stunning sunsets, whitewashed buildings, and breathtaking views over the Aegean Sea.",
-    imgsrc: Santorini,
-  },
-  {
-    id: 3,
-    destination: "Machu Picchu",
-    note: "Embark on a journey to the ancient ruins of the Incas, hidden amidst the Andes mountains, and marvel at the awe-inspiring beauty and rich history.",
-    imgsrc: Machu,
-  },
-  {
-    id: 4,
-    destination: "Bora Bora",
-    note: "Immerse yourself in the allure of this French Polynesian gem, where turquoise lagoons, vibrant coral reefs, and luxurious over-water bungalows await.",
-    imgsrc: Bora,
-  },
-];
-
 export const images = [
   {
     id: 1,
-    imgsrc: Pic1,
+    imgsrc: CTAPic1,
   },
   {
     id: 2,
-    imgsrc: Pic2,
+    imgsrc: CTAPic2,
   },
   {
     id: 3,
-    imgsrc: Pic3,
+    imgsrc: CTAPic3,
   },
   {
     id: 4,
-    imgsrc: Pic4,
+    imgsrc: CTAPic4,
   },
 ];
 
@@ -751,7 +140,7 @@ export const vacationCarousel = [
     subText: "Revitalize Your Mind, Body, and Soul",
     description:
       "Escape to a world of serenity and relaxation with our rejuvenating spa retreats. Unwind amidst soothing ambiance, skilled therapists, and pampering treatments that will leave you feeling refreshed.",
-    imgsrc: Image1,
+    imgsrc: VacationImg1,
   },
   {
     id: 2,
@@ -759,7 +148,7 @@ export const vacationCarousel = [
     subText: "Experience Unmatched Comfort and Adventure",
     description:
       "Embark on a journey of a lifetime aboard our exquisite cruise ships. Enjoy luxurious amenities, breathtaking views, and a wide range of activities for an unforgettable vacation.",
-    imgsrc: Image2,
+    imgsrc: VacationImg2,
   },
   {
     id: 3,
@@ -767,7 +156,7 @@ export const vacationCarousel = [
     subText: "Soar Above the Clouds and Feel Alive",
     description:
       "Experience the ultimate adrenaline rush with our exhilarating skydiving adventures. Dive from high above, feel the rush of wind, and witness awe-inspiring views as you conquer the skies.",
-    imgsrc: Image3,
+    imgsrc: VacationImg3,
   },
   {
     id: 4,
@@ -775,7 +164,7 @@ export const vacationCarousel = [
     subText: "Escape to Tranquility Under the Moonlight",
     description:
       "Immerse yourself in the enchanting beauty of our secluded beaches. Take romantic strolls along the shoreline, bask in the gentle ocean breeze, and witness the mesmerizing spectacle of the night sky.",
-    imgsrc: Image4,
+    imgsrc: VacationImg4,
   },
   {
     id: 5,
@@ -783,7 +172,7 @@ export const vacationCarousel = [
     subText: "Reach New Heights, Unleash Your Adventure",
     description:
       "Embark on a thrilling mountain climbing expedition and conquer majestic peaks. Push your limits, revel in breathtaking vistas, and savor the sense of achievement at the pinnacle of the world.",
-    imgsrc: Image5,
+    imgsrc: VacationImg5,
   },
   {
     id: 6,
@@ -791,7 +180,7 @@ export const vacationCarousel = [
     subText: "Immerse Yourself in Nature's Wonders",
     description:
       "Explore the extraordinary realm of wildlife through an immersive expedition that transports you to nature's wonders. Witness majestic creatures thriving in their natural habitats, from graceful big cats to and elusive marine life.",
-    imgsrc: Image6,
+    imgsrc: VacationImg6,
   },
   {
     id: 7,
@@ -799,7 +188,7 @@ export const vacationCarousel = [
     subText: "Uncover the Pulse of Vibrant Cities",
     description:
       "Indulge in a captivating city escape and immerse yourself in the vibrant energy of bustling metropolises. Wander through historic streets, marvel at iconic landmarks, and discover hidden gems nestled amidst urban landscapes.",
-    imgsrc: Image7,
+    imgsrc: VacationImg7,
   },
   {
     id: 8,
@@ -807,7 +196,7 @@ export const vacationCarousel = [
     subText: "Savor Culinary Excellence",
     description:
       "Elevate your taste buds with an exquisite fine dining experience. Delight in a gourmet journey of flavors, artfully crafted dishes, and impeccable service that will leave you with unforgettable culinary memories.",
-    imgsrc: Image8,
+    imgsrc: VacationImg8,
   },
 ];
 
@@ -818,7 +207,7 @@ export const propertyCarousel = [
     subText: "France, Normandy",
     description:
       "Discover the charm of Normandy in this seaside cottage retreat. Wake up to the sound of waves, explore picturesque coastal towns, and savor fresh seafood. The perfect escape for beach lovers.",
-    imgsrc: Cottage,
+    imgsrc: PropertyImg1,
   },
   {
     id: 2,
@@ -826,7 +215,7 @@ export const propertyCarousel = [
     subText: "Germany, Neuruppin",
     description:
       "Escape to this exclusive oasis in Neuruppin, Germany. Experience serenity and relaxation in a luxurious villa surrounded by beautiful landscapes. Unwind in a private spa, enjoy gourmet dining, and create unforgettable memories.",
-    imgsrc: Villa,
+    imgsrc: PropertyImg2,
   },
   {
     id: 3,
@@ -834,7 +223,7 @@ export const propertyCarousel = [
     subText: "Canada, Banff",
     description:
       "Embrace nature in a cozy mountain view cabin in Banff, Canada. Hike through breathtaking trails, stargaze from your private deck, and experience the beauty of the Canadian Rockies.",
-    imgsrc: Cabin,
+    imgsrc: PropertyImg3,
   },
   {
     id: 4,
@@ -842,7 +231,7 @@ export const propertyCarousel = [
     subText: "USA, Montana",
     description:
       "Unwind in a rustic yet elegant riverside lodge in Montana, USA. Fly fish in pristine waters, go horseback riding through scenic landscapes, and enjoy evenings by the campfire.",
-    imgsrc: Lodge,
+    imgsrc: PropertyImg4,
   },
   {
     id: 5,
@@ -850,7 +239,7 @@ export const propertyCarousel = [
     subText: "Scotland, Edinburgh",
     description:
       "Step back in time with a stay in a historic castle in Edinburgh, Scotland. Explore ancient corridors, enjoy traditional Scottish feasts, and experience the grandeur of a bygone era.",
-    imgsrc: Castle,
+    imgsrc: PropertyImg5,
   },
 ];
 
@@ -859,45 +248,30 @@ export const execs = [
     id: 1,
     name: "Olulanke Mainasara",
     position: "Co-Founder",
-    image: {
-      url: "",
-      alt: "",
-    },
+    image: ExecImg,
   },
   {
     id: 2,
-    name: "Olulanke Oluwatomiwa",
+    name: "Olulanke Tomiwa",
     position: "Co-Founder",
-    image: {
-      url: "",
-      alt: "",
-    },
+    image: ExecImg,
   },
   {
     id: 3,
     name: "Olulanke Folashade",
     position: "C.F.O",
-    image: {
-      url: "",
-      alt: "",
-    },
+    image: ExecImg,
   },
   {
     id: 4,
     name: "Olulanke Isaac",
     position: "C.T.O",
-    image: {
-      url: "",
-      alt: "",
-    },
+    image: ExecImg,
   },
   {
     id: 5,
     name: "Olulanke Ebunoluwa",
     position: "G.C",
-    image: {
-      url: "",
-      alt: "",
-    },
+    image: ExecImg,
   },
 ];
